refactor: make rbac package generic and reusable
